src/test/java/edu/stanford/slac/core_build_system/repository/KubernetesRepositoryTest.java
package edu.stanford.slac.core_build_system.repository;

import io.fabric8.kubernetes.client.KubernetesClient;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.TestInstance;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.DirtiesContext;
import org.springframework.test.context.ActiveProfiles;

import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;

@AutoConfigureMockMvc
@SpringBootTest()
@TestInstance(TestInstance.Lifecycle.PER_CLASS)
@ExtendWith(MockitoExtension.class)
@ActiveProfiles({"test"})
@DirtiesContext(classMode = DirtiesContext.ClassMode.AFTER_CLASS)
public class KubernetesRepositoryTest {
    @Autowired
    KubernetesClient client;
    @Autowired
    KubernetesRepository repository;

    @Test
    public void testK8sPresence() {
        assertDoesNotThrow(()->repository.ensureNamespace("ns1"));
        var result = assertDoesNotThrow(()->repository.existsNamespace("ns1"));
        assertThat(result).isTrue();
    }
}
